+++ client/vscode/src/webview/search-sidebar/AuthSidebarView.tsx
@@ -27,32 +27,37 @@ export const AuthSidebarView: React.FunctionComponent<AuthSidebarViewProps> = ({
     stateStatus,
 }) => {
     const [state, setState] = useState<'initial' | 'validating' | 'success' | 'failure'>('initial')
-
     const [hasAccount, setHasAccount] = useState(false)
-
+    const [usePrivateInstance, setUsePrivateInstance] = useState(false)
     const signUpURL = `https://sourcegraph.com/sign-up?editor=vscode&${SIDEBAR_UTM_PARAMS}`
-
     const instanceHostname = useMemo(() => new URL(instanceURL).hostname, [instanceURL])
+    const [hostname, setHostname] = useState(instanceHostname)
 
     const validateAccessToken: React.FormEventHandler<HTMLFormElement> = (event): void => {
         event.preventDefault()
         if (state !== 'validating') {
+            const newInstanceUrl = (event.currentTarget.elements.namedItem('instance-url') as HTMLInputElement).value
             const newAccessToken = (event.currentTarget.elements.namedItem('token') as HTMLInputElement).value
-
+            let authStateVariables = {
+                request: currentAuthStateQuery,
+                variables: {},
+                mightContainPrivateInfo: true,
+                overrideAccessToken: newAccessToken,
+            }
+            if (usePrivateInstance) {
+                setHostname(newInstanceUrl)
+                authStateVariables = { ...authStateVariables, ...{ overrideSourcegraphURL: newInstanceUrl } }
+            }
             setState('validating')
             const currentAuthStateResult = platformContext
-                .requestGraphQL<CurrentAuthStateResult, CurrentAuthStateVariables>({
-                    request: currentAuthStateQuery,
-                    variables: {},
-                    mightContainPrivateInfo: true,
-                    overrideAccessToken: newAccessToken,
-                })
+                .requestGraphQL<CurrentAuthStateResult, CurrentAuthStateVariables>(authStateVariables)
                 .toPromise()
 
             currentAuthStateResult
-                .then(({ data }) => {
+                .then(async ({ data }) => {
                     if (data?.currentUser) {
                         setState('success')
+                        await extensionCoreAPI.setEndpointUri(newInstanceUrl)
                         return extensionCoreAPI.setAccessToken(newAccessToken)
                     }
                     setState('failure')
@@ -150,7 +155,7 @@ export const AuthSidebarView: React.FunctionComponent<AuthSidebarViewProps> = ({
     return renderCommon(
         <>
             <p className={classNames(styles.ctaParagraph)}>
-                Sign in by entering an access token created through your user settings on {instanceHostname}.
+                Sign in by entering an access token created through your user settings on {hostname}.
             </p>
             <p className={classNames(styles.ctaParagraph)}>
                 See our{' '}
@@ -164,13 +169,15 @@ export const AuthSidebarView: React.FunctionComponent<AuthSidebarViewProps> = ({
             </p>
             {state === 'failure' && (
                 <Alert variant="danger">
-                    Unable to verify your access token for {instanceHostname}. Please try again with a new access token.
+                    Unable to verify your access token for {hostname}. Please try again with a new access token.
                 </Alert>
             )}
             <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
                 <LoaderInput loading={state === 'validating'}>
+                    <label htmlFor="access-token-input">Access Token</label>
                     <input
                         className={classNames('input form-control', styles.ctaInput)}
+                        id="access-token-input"
                         type="text"
                         name="token"
                         required={true}
@@ -181,18 +188,66 @@ export const AuthSidebarView: React.FunctionComponent<AuthSidebarViewProps> = ({
                     />
                 </LoaderInput>
             </p>
-            <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
-                <button
-                    type="submit"
-                    disabled={state === 'validating'}
-                    className={classNames('btn my-1', styles.ctaButton)}
-                >
-                    Enter access token
+            {!usePrivateInstance && (
+                <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
+                    <button
+                        type="submit"
+                        disabled={state === 'validating'}
+                        className={classNames('btn my-1', styles.ctaButton)}
+                    >
+                        Authenticate Access
+                    </button>
+                </p>
+            )}
+            {!usePrivateInstance ? (
+                <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
+                    <button type="button" className="btn btn-text-link h-0" onClick={() => setUsePrivateInstance(true)}>
+                        Need to connect to a private instance?
+                    </button>
+                </p>
+            ) : (
+                <>
+                    <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
+                        <LoaderInput loading={state === 'validating'}>
+                            <label htmlFor="private-url-input">Sourcegraph Instance URL</label>
+                            <input
+                                className={classNames('input form-control', styles.ctaInput)}
+                                id="private-url-input"
+                                type="text"
+                                name="instance-url"
+                                required={true}
+                                autoFocus={true}
+                                spellCheck={false}
+                                disabled={state === 'validating'}
+                                placeholder="ex sourcegraph.example.com"
+                            />
+                        </LoaderInput>
+                    </p>
+                    <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
+                        <button
+                            type="submit"
+                            disabled={state === 'validating'}
+                            className={classNames('btn my-1', styles.ctaButton)}
+                        >
+                            Authenticate Access
+                        </button>
+                    </p>
+                    <p className={classNames(styles.ctaButtonWrapperWithContextBelow)}>
+                        <button
+                            type="button"
+                            className="btn btn-text-link h-0"
+                            onClick={() => setUsePrivateInstance(false)}
+                        >
+                            Not a private instance user?
+                        </button>
+                    </p>
+                </>
+            )}
+            <p className={classNames(styles.ctaParagraph)}>
+                <button type="button" className="btn btn-text-link h-0" onClick={onSignUpClick}>
+                    Create an account
                 </button>
             </p>
-            <a href="/" className={classNames(styles.ctaParagraph)} onClick={onSignUpClick}>
-                Create an account
-            </a>
         </>
     )
 }